requirements.txt
Flask~=2.2.5
PyMySQL~=1.1.0
DBUtils~=3.0.3
flask_cors~=4.0.0
flask_socketio~=5.3.6
gevent~=23.9.1
PyJWT~=2.8.0